style(branches): share the sidebar's project row with the move picker
The move dialog's project rows showed a plain name over full repo-path
labels, so the same projects read differently than they do in the
sidebar, which introduces each one with a state icon and colored repo
badges. The sidebar row's identity block — the PR-status or cloud icon
on the left, the bold name, and the badge/activity meta line with its
repo-count fallback — now lives in a shared ProjectRowContent component
that both the sidebar and MoveBranchDialog render. It derives everything
from the shared stores given only the project, so a picker row shows the
same live state the sidebar does.

The dialog needed nothing new to feed it: opening it already hydrates
every project's branches and repos for the duplicate-repo check, and
hydration is also what materializes the repo badges.

The sidebar's active-row brightening reached its meta text through
scoped descendant selectors, which can't cross into the child, and a
:global svg override would tie with the child's status-icon colors at
equal specificity, leaving the winner to stylesheet order. The meta
color is now a --project-row-meta-color custom property the child
reads, and the stroke override names the two places that still need it
— the row's status spinners and the All Repos icon — where its higher
specificity wins deterministically.

Verified with prettier, svelte-check (0 errors, 0 warnings) and 648
frontend tests.
